+++ crates/canic/src/macros/endpoints/bundles.rs
@@ -13,6 +13,7 @@
 macro_rules! canic_bundle_shared_runtime_endpoints {
     () => {
         $crate::canic_emit_lifecycle_core_endpoints!();
+        #[cfg(canic_memory_ledger_enabled)]
         $crate::canic_emit_memory_ledger_diagnostic_endpoint!();
         $crate::canic_bundle_discovery_endpoints!();
         $crate::canic_bundle_observability_endpoints!();
@@ -51,6 +52,7 @@ macro_rules! canic_bundle_nonroot_only_endpoints {
 macro_rules! canic_bundle_wasm_store_runtime_endpoints {
     () => {
         $crate::canic_emit_lifecycle_core_endpoints!();
+        #[cfg(canic_memory_ledger_enabled)]
         $crate::canic_emit_memory_ledger_diagnostic_endpoint!();
         $crate::canic_bundle_discovery_endpoints!();
         #[cfg(not(canic_disable_bundle_cycle_tracker))]